+++ docs/changelog.rst
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-0.8.0 (Unreleased)
-------------------
+Unreleased
+----------
 
 Changed
 ~~~~~~~